requirements.txt
# python 3.10.14

# Search-TTA
scipy==1.14.1
scikit-learn==1.6.1
scikit-image==0.24.0
matplotlib==3.9.1
imageio==2.36.0
shapely==2.0.7
rasterio==1.4.1
kneed==0.8.5
easydict==1.13
ray==2.6.1
openpyxl==3.1.5
tensorboard==2.18.0

# Taxabind-AVS
numpy==1.26.3
torch==2.4.1 
torchvision==0.19.1 
torchaudio==2.4.1 
pytorch-lightning==2.2.1
open_clip_torch==2.30.0
transformers==4.45.1
tokenizers==0.20.3
opencv-python==4.10.0.84
gradio==4.44.1 
datasets==3.1.0